+++ pkg/glib/glib.go
@@ -5,28 +5,29 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
-	"log"
+	"strconv"
 
 	"github.com/kubermatic-labs/gman/pkg/config"
 	password "github.com/sethvargo/go-password/password"
 	"golang.org/x/oauth2/google"
 	admin "google.golang.org/api/admin/directory/v1"
+	groupssettings "google.golang.org/api/groupssettings/v1"
 	"google.golang.org/api/option"
 )
 
-// CreateDirectoryService() creates a client for communicating with Google APIs,
-// returns an Admin SDK Directory service object authorized with.
+// NewDirectoryService() creates a client for communicating with Google Directory API,
+// returns a service object authorized to perform actions in Gsuite.
 func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) (*admin.Service, error) {
 	ctx := context.Background()
 
 	jsonCredentials, err := ioutil.ReadFile(clientSecretFile)
 	if err != nil {
-		return nil, fmt.Errorf("ReadFile(clientSecretFile): %v", err)
+		return nil, fmt.Errorf("unable to read json credentials (clientSecretFile): %v", err)
 	}
 
-	config, err := google.JWTConfigFromJSON(jsonCredentials, admin.AdminDirectoryUserScope, admin.AdminDirectoryGroupScope, admin.AdminDirectoryGroupMemberScope, admin.AdminDirectoryOrgunitScope)
+	config, err := google.JWTConfigFromJSON(jsonCredentials, admin.AdminDirectoryUserScope, admin.AdminDirectoryGroupScope, admin.AdminDirectoryGroupMemberScope, admin.AdminDirectoryOrgunitScope, admin.AdminDirectoryResourceCalendarScope)
 	if err != nil {
-		return nil, fmt.Errorf("JWTConfigFromJSON: %v", err)
+		return nil, fmt.Errorf("unable to process credentials: %v", err)
 	}
 	config.Subject = impersonatedUserEmail
 
@@ -34,7 +35,32 @@ func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) 
 
 	srv, err := admin.NewService(ctx, option.WithTokenSource(ts))
 	if err != nil {
-		return nil, fmt.Errorf("NewService: %v", err)
+		return nil, fmt.Errorf("unable to create a new Admin Service: %v", err)
+	}
+	return srv, nil
+}
+
+// NewGroupsService() creates a client for communicating with Google Groupssettings API,
+// returns a service object authorized to perform actions in Gsuite.
+func NewGroupsService(clientSecretFile string, impersonatedUserEmail string) (*groupssettings.Service, error) {
+	ctx := context.Background()
+
+	jsonCredentials, err := ioutil.ReadFile(clientSecretFile)
+	if err != nil {
+		return nil, fmt.Errorf("unable to read json credentials (clientSecretFile): %v", err)
+	}
+
+	config, err := google.JWTConfigFromJSON(jsonCredentials, groupssettings.AppsGroupsSettingsScope)
+	if err != nil {
+		return nil, fmt.Errorf("unable to process credentials: %v", err)
+	}
+	config.Subject = impersonatedUserEmail
+
+	ts := config.TokenSource(ctx)
+
+	srv, err := groupssettings.NewService(ctx, option.WithTokenSource(ts))
+	if err != nil {
+		return nil, fmt.Errorf("unable to create a new Groupssettings Service: %v", err)
 	}
 	return srv, nil
 }
@@ -47,8 +73,7 @@ func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) 
 func GetListOfUsers(srv admin.Service) ([]*admin.User, error) {
 	request, err := srv.Users.List().Customer("my_customer").OrderBy("email").Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve users in domain: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("unable to retrieve list of users in domain: %v", err)
 	}
 	return request.Users, nil
 }
@@ -73,18 +98,19 @@ func CreateUser(srv admin.Service, user *config.UserConfig) error {
 	// generate a rand password
 	pass, err := password.Generate(20, 5, 5, false, false)
 	if err != nil {
-		log.Fatalf("Unable to generate password: %v", err)
-		return err
+		return fmt.Errorf("unable to generate password: %v", err)
 	}
-	newUser := createGSuiteUserFromConfig(user)
+	newUser := createGSuiteUserFromConfig(srv, user)
 	newUser.Password = pass
 	newUser.ChangePasswordAtNextLogin = true
 
 	_, err = srv.Users.Insert(newUser).Do()
 	if err != nil {
-		log.Fatalf("Unable to create a user: %v", err)
-		return err
+		return fmt.Errorf("unable to insert a user: %v", err)
 	}
+
+	HandleUserAliases(srv, newUser, user.Aliases)
+
 	return nil
 }
 
@@ -92,25 +118,84 @@ func CreateUser(srv admin.Service, user *config.UserConfig) error {
 func DeleteUser(srv admin.Service, user *admin.User) error {
 	err := srv.Users.Delete(user.PrimaryEmail).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a user: %v", err)
-		return err
+		return fmt.Errorf("unable to delete a user %s: %v", user.PrimaryEmail, err)
 	}
 	return nil
 }
 
 // UpdateUser updates the remote user with config
 func UpdateUser(srv admin.Service, user *config.UserConfig) error {
-	updatedUser := createGSuiteUserFromConfig(user)
+	updatedUser := createGSuiteUserFromConfig(srv, user)
 	_, err := srv.Users.Update(user.PrimaryEmail, updatedUser).Do()
 	if err != nil {
-		log.Fatalf("Unable to update a user: %v", err)
-		return err
+		return fmt.Errorf("unable to update a user %s: %v", user.PrimaryEmail, err)
 	}
+
+	HandleUserAliases(srv, updatedUser, user.Aliases)
+
+	return nil
+}
+
+func HandleUserAliases(srv admin.Service, googleUser *admin.User, configAliases []string) error {
+	request, err := srv.Users.Aliases.List(googleUser.PrimaryEmail).Do()
+	if err != nil {
+		return fmt.Errorf("unable to list user aliases in GSuite: %v", err)
+	}
+
+	if len(configAliases) == 0 {
+		for _, alias := range request.Aliases {
+			err = srv.Users.Aliases.Delete(googleUser.PrimaryEmail, fmt.Sprint(alias.(map[string]interface{})["alias"])).Do()
+			if err != nil {
+				return fmt.Errorf("unable to delete user alias: %v", err)
+			}
+		}
+	} else {
+		// check aliases to delete
+		for _, alias := range request.Aliases {
+			found := false
+			for _, configAlias := range configAliases {
+				if alias.(map[string]interface{})["alias"] == configAlias {
+					found = true
+					break
+				}
+			}
+			if !found {
+				// delete
+				err = srv.Users.Aliases.Delete(googleUser.PrimaryEmail, fmt.Sprint(alias.(map[string]interface{})["alias"])).Do()
+				if err != nil {
+					return fmt.Errorf("unable to delete user alias: %v", err)
+				}
+			}
+
+		}
+	}
+
+	// check aliases to add
+	for _, configAlias := range configAliases {
+		found := false
+		for _, alias := range request.Aliases {
+			if alias.(map[string]interface{})["alias"] == configAlias {
+				found = true
+				break
+			}
+		}
+		if !found {
+			// add
+			newAlias := &admin.Alias{
+				Alias: configAlias,
+			}
+			_, err = srv.Users.Aliases.Insert(googleUser.PrimaryEmail, newAlias).Do()
+			if err != nil {
+				return fmt.Errorf("unable to add user alias: %v", err)
+			}
+		}
+	}
+
 	return nil
 }
 
 // createGSuiteUserFromConfig converts a ConfigUser to (GSuite) admin.User
-func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
+func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *admin.User {
 	googleUser := &admin.User{
 		Name: &admin.UserName{
 			GivenName:  user.FirstName,
@@ -120,9 +205,38 @@ func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
 		OrgUnitPath:  user.OrgUnitPath,
 	}
 
+	if len(user.Phones) > 0 {
+		phNums := []admin.UserPhone{}
+		for _, phone := range user.Phones {
+			phNum := admin.UserPhone{
+				Value: phone,
+				Type:  "home",
+			}
+			phNums = append(phNums, phNum)
+		}
+		googleUser.Phones = phNums
+	}
+
+	if user.Address != "" {
+		addr := []admin.UserAddress{
+			{
+				Formatted: user.Address,
+				Type:      "home",
+			},
+		}
+		googleUser.Addresses = addr
+	}
+
+	if user.RecoveryEmail != "" {
+		googleUser.RecoveryEmail = user.RecoveryEmail
+	}
+
+	if user.RecoveryPhone != "" {
+		googleUser.RecoveryPhone = user.RecoveryPhone
+	}
+
 	if user.SecondaryEmail != "" {
-		googleUser.RecoveryEmail = user.SecondaryEmail
-		workEm := &[]admin.UserEmail{
+		workEm := []admin.UserEmail{
 			{
 				Address: user.SecondaryEmail,
 				Type:    "work",
@@ -131,7 +245,119 @@ func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
 		googleUser.Emails = workEm
 	}
 
+	if user.Employee != (config.EmployeeConfig{}) {
+		uOrg := []admin.UserOrganization{
+			{
+				Department:  user.Employee.Department,
+				Title:       user.Employee.JobTitle,
+				CostCenter:  user.Employee.CostCenter,
+				Description: user.Employee.Type,
+			},
+		}
+		googleUser.Organizations = uOrg
+		rel := []admin.UserRelation{
+			{
+				Value: user.Employee.ManagerEmail,
+				Type:  "manager",
+			},
+		}
+		googleUser.Relations = rel
+
+		ids := []admin.UserExternalId{
+			{
+				Value: user.Employee.EmployeeID,
+				Type:  "organization",
+			},
+		}
+		googleUser.ExternalIds = ids
+
+	}
+
+	if user.Location != (config.LocationConfig{}) {
+		loc := []admin.UserLocation{
+			{
+				Area:         "desk",
+				BuildingId:   user.Location.Building,
+				FloorName:    user.Location.Floor,
+				FloorSection: user.Location.FloorSection,
+				Type:         "desk",
+			},
+		}
+		googleUser.Locations = loc
+
+	}
+
 	return googleUser
+}
+
+// createConfigUserFromGSuite converts a (GSuite) admin.User to ConfigUser
+func CreateConfigUserFromGSuite(googleUser *admin.User) config.UserConfig {
+	// get emails
+	primaryEmail, secondaryEmail := GetUserEmails(googleUser)
+
+	configUser := config.UserConfig{
+		FirstName:      googleUser.Name.GivenName,
+		LastName:       googleUser.Name.FamilyName,
+		PrimaryEmail:   primaryEmail,
+		SecondaryEmail: secondaryEmail,
+		OrgUnitPath:    googleUser.OrgUnitPath,
+		RecoveryPhone:  googleUser.RecoveryPhone,
+		RecoveryEmail:  googleUser.RecoveryEmail,
+	}
+
+	if len(googleUser.Aliases) > 0 {
+		for _, alias := range googleUser.Aliases {
+			configUser.Aliases = append(configUser.Aliases, string(alias))
+		}
+	}
+
+	if googleUser.Phones != nil {
+		for _, phone := range googleUser.Phones.([]interface{}) {
+			configUser.Phones = append(configUser.Phones, fmt.Sprint(phone.(map[string]interface{})["value"]))
+		}
+	}
+
+	if googleUser.ExternalIds != nil {
+		for _, id := range googleUser.ExternalIds.([]interface{}) {
+			if id.(map[string]interface{})["type"] == "organization" {
+				configUser.Employee.EmployeeID = fmt.Sprint(id.(map[string]interface{})["value"])
+			}
+		}
+	}
+
+	if googleUser.Organizations != nil {
+		for _, org := range googleUser.Organizations.([]interface{}) {
+			configUser.Employee.Department = fmt.Sprint(org.(map[string]interface{})["department"])
+			configUser.Employee.JobTitle = fmt.Sprint(org.(map[string]interface{})["title"])
+			configUser.Employee.Type = fmt.Sprint(org.(map[string]interface{})["description"])
+			configUser.Employee.CostCenter = fmt.Sprint(org.(map[string]interface{})["costCenter"])
+		}
+	}
+
+	if googleUser.Relations != nil {
+		for _, rel := range googleUser.Relations.([]interface{}) {
+			if rel.(map[string]interface{})["type"] == "manager" {
+				configUser.Employee.ManagerEmail = fmt.Sprint(rel.(map[string]interface{})["value"])
+			}
+		}
+	}
+
+	if googleUser.Locations != nil {
+		for _, loc := range googleUser.Locations.([]interface{}) {
+			configUser.Location.Building = fmt.Sprint(loc.(map[string]interface{})["buildingId"])
+			configUser.Location.Floor = fmt.Sprint(loc.(map[string]interface{})["floorName"])
+			configUser.Location.FloorSection = fmt.Sprint(loc.(map[string]interface{})["floorSection"])
+		}
+	}
+	if googleUser.Addresses != nil {
+		for _, addr := range googleUser.Addresses.([]interface{}) {
+			if addr.(map[string]interface{})["type"] == "home" {
+				configUser.Address = fmt.Sprint(addr.(map[string]interface{})["formatted"])
+			}
+		}
+	}
+
+	return configUser
 }
 
 //----------------------------------------//
@@ -142,23 +368,35 @@ func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
 func GetListOfGroups(srv *admin.Service) ([]*admin.Group, error) {
 	request, err := srv.Groups.List().Customer("my_customer").OrderBy("email").Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve groups in domain: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("unable to retrieve a list of groups in domain: %v", err)
 	}
 	return request.Groups, nil
 }
 
+// GetSettingOfGroup returns a group settings object from the API
+func GetSettingOfGroup(srv *groupssettings.Service, groupId string) (*groupssettings.Groups, error) {
+	request, err := srv.Groups.Get(groupId).Do()
+	if err != nil {
+		return nil, fmt.Errorf("unable to retrieve group's (%s) settings: %v", groupId, err)
+	}
+	return request, nil
+}
+
 // CreateGroup creates a new group in GSuite via their API
-func CreateGroup(srv admin.Service, group *config.GroupConfig) error {
-	newGroup := createGSuiteGroupFromConfig(group)
+func CreateGroup(srv admin.Service, grSrv groupssettings.Service, group *config.GroupConfig) error {
+	newGroup, groupSettings := CreateGSuiteGroupFromConfig(group)
 	_, err := srv.Groups.Insert(newGroup).Do()
 	if err != nil {
-		log.Fatalf("Unable to create a group: %v", err)
-		return err
+		return fmt.Errorf("unable to insert a group: %v", err)
 	}
 	// add the members
 	for _, member := range group.Members {
 		AddNewMember(srv, newGroup.Email, &member)
+	}
+	// add the group's settings
+	_, err = grSrv.Groups.Update(newGroup.Email, groupSettings).Do()
+	if err != nil {
+		return fmt.Errorf("unable to set the group settings: %v", err)
 	}
 	return nil
 }
@@ -167,25 +405,29 @@ func CreateGroup(srv admin.Service, group *config.GroupConfig) error {
 func DeleteGroup(srv admin.Service, group *admin.Group) error {
 	err := srv.Groups.Delete(group.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a group: %v", err)
-		return err
+		return fmt.Errorf("unable to delete a group: %v", err)
 	}
 	return nil
 }
 
 // UpdateGroup updates the remote group with config
-func UpdateGroup(srv admin.Service, group *config.GroupConfig) error {
-	updatedGroup := createGSuiteGroupFromConfig(group)
+func UpdateGroup(srv admin.Service, grSrv groupssettings.Service, group *config.GroupConfig) error {
+	updatedGroup, groupSettings := CreateGSuiteGroupFromConfig(group)
 	_, err := srv.Groups.Update(group.Email, updatedGroup).Do()
 	if err != nil {
-		log.Fatalf("Unable to update a group: %v", err)
-		return err
+		return fmt.Errorf("unable to update a group: %v", err)
 	}
+	// update group's settings
+	_, err = grSrv.Groups.Update(group.Email, groupSettings).Do()
+	if err != nil {
+		return fmt.Errorf("unable to update group settings: %v", err)
+	}
+
 	return nil
 }
 
 // createGSuiteGroupFromConfig converts a ConfigGroup to (GSuite) admin.Group
-func createGSuiteGroupFromConfig(group *config.GroupConfig) *admin.Group {
+func CreateGSuiteGroupFromConfig(group *config.GroupConfig) (*admin.Group, *groupssettings.Groups) {
 	googleGroup := &admin.Group{
 		Name:  group.Name,
 		Email: group.Email,
@@ -193,7 +435,55 @@ func createGSuiteGroupFromConfig(group *config.GroupConfig) *admin.Group {
 	if group.Description != "" {
 		googleGroup.Description = group.Description
 	}
-	return googleGroup
+
+	groupSettings := &groupssettings.Groups{
+		WhoCanContactOwner:   group.WhoCanContactOwner,
+		WhoCanViewMembership: group.WhoCanViewMembership,
+		WhoCanApproveMembers: group.WhoCanApproveMembers,
+		WhoCanPostMessage:    group.WhoCanPostMessage,
+		WhoCanJoin:           group.WhoCanJoin,
+		IsArchived:           strconv.FormatBool(group.IsArchived),
+		ArchiveOnly:          strconv.FormatBool(group.IsArchived),
+		AllowExternalMembers: strconv.FormatBool(group.AllowExternalMembers),
+	}
+
+	return googleGroup, groupSettings
+}
+
+func CreateConfigGroupFromGSuite(googleGroup *admin.Group, members []*admin.Member, gSettings *groupssettings.Groups) (config.GroupConfig, error) {
+
+	boolAllowExternalMembers, err := strconv.ParseBool(gSettings.AllowExternalMembers)
+	if err != nil {
+		return config.GroupConfig{}, fmt.Errorf("could not parse 'AllowExternalMembers' value from string to bool: %v", err)
+	}
+	boolIsArchived, err := strconv.ParseBool(gSettings.IsArchived)
+	if err != nil {
+		return config.GroupConfig{}, fmt.Errorf("could not parse 'IsArchived' value from string to bool: %v", err)
+	}
+
+	configGroup := config.GroupConfig{
+		Name:                 googleGroup.Name,
+		Email:                googleGroup.Email,
+		Description:          googleGroup.Description,
+		WhoCanContactOwner:   gSettings.WhoCanContactOwner,
+		WhoCanViewMembership: gSettings.WhoCanViewMembership,
+		WhoCanApproveMembers: gSettings.WhoCanApproveMembers,
+		WhoCanPostMessage:    gSettings.WhoCanPostMessage,
+		WhoCanJoin:           gSettings.WhoCanJoin,
+		AllowExternalMembers: boolAllowExternalMembers,
+		IsArchived:           boolIsArchived,
+		Members:              []config.MemberConfig{},
+	}
+
+	for _, m := range members {
+		configGroup.Members = append(configGroup.Members, config.MemberConfig{
+			Email: m.Email,
+			Role:  m.Role,
+		})
+
+	}
+
+	return configGroup, nil
 }
 
 //----------------------------------------//
@@ -204,8 +494,7 @@ func createGSuiteGroupFromConfig(group *config.GroupConfig) *admin.Group {
 func GetListOfMembers(srv *admin.Service, group *admin.Group) ([]*admin.Member, error) {
 	request, err := srv.Members.List(group.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve members in group %s: %v", group.Name, err)
-		return nil, err
+		return nil, fmt.Errorf("unable to retrieve members in group %s: %v", group.Name, err)
 	}
 	return request.Members, nil
 }
@@ -215,8 +504,7 @@ func AddNewMember(srv admin.Service, groupEmail string, member *config.MemberCon
 	newMember := createGSuiteGroupMemberFromConfig(member)
 	_, err := srv.Members.Insert(groupEmail, newMember).Do()
 	if err != nil {
-		log.Fatalf("Unable to add a member: %v", err)
-		return err
+		return fmt.Errorf("unable to add a member to a group: %v", err)
 	}
 	return nil
 }
@@ -225,20 +513,18 @@ func AddNewMember(srv admin.Service, groupEmail string, member *config.MemberCon
 func RemoveMember(srv admin.Service, groupEmail string, member *admin.Member) error {
 	err := srv.Members.Delete(groupEmail, member.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a member: %v", err)
-		return err
+		return fmt.Errorf("unable to delete a member from a group: %v", err)
 	}
 	return nil
 }
 
 // MemberExists checks if member exists in group
-func MemberExists(srv admin.Service, group *admin.Group, member *config.MemberConfig) bool {
+func MemberExists(srv admin.Service, group *admin.Group, member *config.MemberConfig) (bool, error) {
 	exists, err := srv.Members.HasMember(group.Email, member.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to check if member %s exists in a group %s: %v", member.Email, group.Name, err)
-		return false
+		return false, fmt.Errorf("unable to check if member %s exists in a group %s: %v", member.Email, group.Name, err)
 	}
-	return exists.IsMember
+	return exists.IsMember, nil
 }
 
 // UpdateMembership changes the role of the member
@@ -247,8 +533,7 @@ func UpdateMembership(srv admin.Service, groupEmail string, member *config.Membe
 	newMember := createGSuiteGroupMemberFromConfig(member)
 	_, err := srv.Members.Update(groupEmail, member.Email, newMember).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a member: %v", err)
-		return err
+		return fmt.Errorf("unable to update a member in a group: %v", err)
 	}
 	return nil
 }
@@ -270,8 +555,7 @@ func createGSuiteGroupMemberFromConfig(member *config.MemberConfig) *admin.Membe
 func GetListOfOrgUnits(srv *admin.Service) ([]*admin.OrgUnit, error) {
 	request, err := srv.Orgunits.List("my_customer").Type("all").Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve OrgUnits in domain: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("unable to list OrgUnits in domain: %v", err)
 	}
 	return request.OrganizationUnits, nil
 }
@@ -281,8 +565,7 @@ func CreateOrgUnit(srv admin.Service, ou *config.OrgUnitConfig) error {
 	newOU := createGSuiteOUFromConfig(ou)
 	_, err := srv.Orgunits.Insert("my_customer", newOU).Do()
 	if err != nil {
-		log.Fatalf("Unable to create an org unit: %v", err)
-		return err
+		return fmt.Errorf("unable to create an org unit: %v", err)
 	}
 	return nil
 }
@@ -299,8 +582,7 @@ func DeleteOrgUnit(srv admin.Service, ou *admin.OrgUnit) error {
 
 	err := srv.Orgunits.Delete("my_customer", orgUPath).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete an org unit: %v", err)
-		return err
+		return fmt.Errorf("unable to delete an org unit: %v", err)
 	}
 	return nil
 }
@@ -318,8 +600,7 @@ func UpdateOrgUnit(srv admin.Service, ou *config.OrgUnitConfig) error {
 
 	_, err := srv.Orgunits.Update("my_customer", orgUPath, updatedOu).Do()
 	if err != nil {
-		log.Fatalf("Unable to update an org unit: %v", err)
-		return err
+		return fmt.Errorf("unable to update an org unit: %v", err)
 	}
 	return nil
 }